Added files via upload
First wave of changes made. The candidate frequencies are specific to the guitar. A hashmap with notes and their respective frequencies has been made. Frequencies detected and notes can now be displayed ( e.g. C4). The time-domain plot has been disabled for freeing up resources. Attempts are being made to associate notes with strings and frets in the functions file, but are currently unsuccessful due to it being too computationally heavy with 3 nested for loops.