Added files via upload
Patch 2 implements a BST for the f0 candidates. Each node is unique to one candidate, and each node contains an array for each position on the fretboard that frequency can be played at. BTNode and BST classes were added to do this. The assignment of note positions is done in the Functions class. The Klapuri class now puts the F0s in an array. The array then searches (only the lowest frequency found between 82 Hz and 1047 Hz is used) the BST, and the positions for the note is returned. Next step is to find the best position for the detected note using previously detected notes.