Added files via upload
Current build listens to user, looks for frequencies on the fretboard, and saves them to a vector to be used when the user is done playing. If a frequency that is detected is not on the fretboard, a placeholder frequency is added to denote nothing was played to keep in time with what was played.